requirements.txt
annotated-types==0.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
boto3==1.34.78
botocore==1.34.78
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.8.0rc2
Django==5.0.4
django-cors-headers==4.3.1
django-sendgrid-v5==1.2.3
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
fonttools==4.51.0
idna==3.6
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.1
sendgrid==6.11.0
sendgrid-django==4.2.0
setuptools==65.5.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
